services/backend/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
anthropic>=0.40.0
openai>=1.57.0
asyncpg>=0.30.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
httpx>=0.27.0
python-telegram-bot>=22.0
discord.py>=2.4.0
